backv2-github/DevDocs/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pymupdf>=1.19.0
pytesseract>=0.3.8
opencv-python>=4.5.3
pillow>=8.3.1
google-cloud-secret-manager>=2.12.0
google-generativeai>=0.3.0

# PDF processing
camelot-py>=0.9.0
pdfminer.six>=20201018
pdf2image>=1.16.0
reportlab>=3.6.1
fpdf>=1.7.2

# Data processing and visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1
tabulate>=0.8.9

# OCR and image processing
easyocr>=1.4.1
pytesseract>=0.3.8
opencv-python>=4.5.3

# NLP and ML
transformers>=4.11.3
scikit-learn>=1.0.0
nltk>=3.6.3

# Web and API
requests>=2.26.0
beautifulsoup4>=4.10.0
fastapi>=0.70.0
uvicorn>=0.15.0
